Add files via upload
This is the functionalvery Basic Calculator, which is only performs Addition, Substraction, Multiplication and Division. it is developed by using HTML for definging Structure of the Web page, CSS for definging Styles which improves look and feel of the web page and JavaScript for adding interactiity and dynamic behaviour towards the Web page.